requirements.txt
pandas~=1.3.4
sqlalchemy~=1.3.5
pymysql~=0.9.3
tushare~=1.2.45
mysqlclient~=1.4.4
requests~=2.22.0
wxpy~=0.3.9.8
numpy~=1.16.4
matplotlib~=3.4.3
partd~=1.0.0
python-dateutil~=2.8.0
scikit-learn~=0.21.2
jieba~=0.39
gensim~=3.8.0
lxml~=4.3.4
bs4~=0.0.1
beautifulsoup4~=4.10.0
django~=2.2.5
wordcloud~=1.5.0
timedelta~=2019.4.13
zmail~=0.2.8
schedule~=0.6.0
scipy~=1.2.1
seaborn~=0.9.0
pyecharts~=1.5.1
tqdm~=4.62.3
pytest~=5.0.1
akshare~=1.1.82